[agent_builder] - refactor: streamline defaults handling for knowledge capabilities
- Removed the `getKnowledgeDefaultValues` and  `getInitialPageId` functions from `KnowledgeConfigurationSheet` in favor of utils
- Eliminated redundant transformation code by utilizing existing utility functions for data tree conversion
- Simplified Knowledge configuration code by taking advantage of utility functions for initial values
- Adjusted the padding for `MCPServerSelectionPage` to enhance user interface consistency
